+++ src/extension.ts
@@ -1,6 +1,8 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
+import {createHelloWorldPython} from './pycommand';
+import { getOS } from './checkDevice';
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -16,10 +18,28 @@ export function activate(context: vscode.ExtensionContext) {
 	const disposable = vscode.commands.registerCommand('hallo-wereld.helloWorld', () => {
 		// The code you place here will be executed every time your command is executed
 		// Display a message box to the user
-		vscode.window.showInformationMessage('Hello World from Hallo-Wereld!');
+		vscode.window.showInformationMessage('Hello World from vs-code');
 	});
 
 	context.subscriptions.push(disposable);
+
+	const python = vscode.commands.registerCommand('hallo-world.HelloWorldPython', async () => {
+		const device = await getOS();
+		switch (device) {
+			case 'Windows':
+				createHelloWorldPython('python');
+				break;
+			case 'Mac':
+			case 'Linux':
+				createHelloWorldPython('python3');
+				break;
+			default:
+				vscode.window.showErrorMessage(`${process.platform} is not supported`);
+				break;
+    	}
+	});
+
+	context.subscriptions.push(python);
 }
 
 // This method is called when your extension is deactivated