+++ CHANGELOG.md
@@ -6,4 +6,21 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Added
+- V1.0 Intial commit
+- V1.1 Created a command new command
+- V1.2 Created a function that checks the workspace
+- V1.3 If workspace not found, stop the command
+- V1.4 Create a python file with hello world in it
+- V1.5 Changed the file structure
+- V1.6 Made a check for extensions
+- V1.7 Made a choice for the users to add/enable the missing extensions
+- V1.8 Made the logic that enables/installs the extensions
+- V1.9 {BUG} .venv is created, but the interpreter is not selected correctly
+- V1.10 bugfix .venv works correct now
+- V1.11 pythoncheck / made filestructure cleaner
+- V1.12 pythoncheck now made available on mac and linux, code is also cleaner
+- V1.13 create .venv now also available on mac and linux, code is also cleaner
+- V1.14 {BUG}, code does not work on linux/mac
+
 - Initial release